BUG: handle integer inputs in LLE (fixes #6033) (#6282)

+++ sklearn/manifold/locally_linear.py
@@ -626,7 +626,7 @@ class LocallyLinearEmbedding(BaseEstimator, TransformerMixin):
                                       n_jobs=self.n_jobs)
 
         random_state = check_random_state(self.random_state)
-        X = check_array(X)
+        X = check_array(X, dtype=float)
         self.nbrs_.fit(X)
         self.embedding_, self.reconstruction_error_ = \
             locally_linear_embedding(
